+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## unreleased
+
+### Added
+
+- Enable AuthenticationEvents index,view,add actions through REST API.
+- Support "User login" authentication event type.
+
+### Fixed
+
+- Broken filtering of Authentication Events using the identifier
+
 ## [3.2.5-rciam] - 2021-09-27
 
 ### Added